feat(memory) add purge and get_all_to_prune_events functions
Added fully functional purge function, a function that retrieves all events that are older than certain threshold for further pruning.

+++ api/schemas.py
@@ -10,5 +10,5 @@ class GameEventModel(BaseModel):
     label: str
     timestamp: datetime
     location: Optional[str] = "Not assigned."
-    importance: Optional[int] = 1
+    importance: Optional[float] = 1
     details: Optional[Dict[str, Any]] = None
